Remote Sensing Control
The program is based on the UDP data transmission protocol and can realize the control of the orientation movement of another ESP32-based car using a handheld IMU (MPU6050 in this program) attitude sensor (tied to the ESP32). The specific direction of movement can be adjusted according to actual conditions.